feat: Add Notification entity
This commit adds the Notification entity to the project. The Notification entity represents a notification that can be sent to users. It includes properties such as id, createdAt, updatedAt, message, title, isRead, readAt, isDeleted, deletedAt, type, priority, link, and user. This entity will be used to store and manage notifications in the system.